add check history

+++ GameOnePage.cpp
@@ -15,6 +15,7 @@ GameOnePage::GameOnePage(QWidget *widget,int gameNumber)
     resumeGame = new QPushButton("Resume old game");
     back = new QPushButton("Back");
     description = new QPushButton("Description");
+    checkHistory = new QPushButton("Check your history");
 
     selectLevel = new QComboBox();
     selectLevel->addItem("Select Level");
@@ -37,16 +38,23 @@ GameOnePage::GameOnePage(QWidget *widget,int gameNumber)
     nameL= new QLabel("     username"); // user actual username KEEP spaces
 
     setVerticalLayout();
-    widget->setFixedSize(500,300);
+    widget->setFixedSize(500,350);
     widget->setLayout(verticalLayout);
 
     QObject::connect(back, SIGNAL(clicked()), SLOT(backClicked()));
     QObject::connect(description, SIGNAL(clicked()), SLOT(descriptionClicked()));
 
 }
+
 void GameOnePage::descriptionClicked(){
-    Message *msg = new Message("SpongeBob has to move around the screen\nto collect items that increase his immunity level\nin order to be able to kill more bacteria\nand thus clean the aquarium.");
-    msg->show();
+    if(gameNumber==1){
+        Message *msg = new Message("SpongeBob has to move around the screen\nto collect items that increase his immunity level\nin order to be able to kill more bacteria\nand thus clean the aquarium.");
+        msg->show();
+    }
+    else{
+        Message *msg = new Message("Game 2 is still in the unknown :)!");
+        msg->show();
+    }
 }
 
 void GameOnePage::backClicked(){
@@ -68,5 +76,6 @@ void GameOnePage::setVerticalLayout()
     verticalLayout->addWidget(newGame);
     verticalLayout->addWidget(resumeGame);
     verticalLayout->addWidget(selectLevel);
+    verticalLayout->addWidget(checkHistory);
     verticalLayout->addWidget(back);
 }
